20090203 jhcのprotocol_moduleコードマージ
git-svn-id: http://10.144.169.20/repos/um/branches/l7vsd-3.x-shamshel@6574 1ed66053-1c2d-0410-8867-f7571e6e31d3

+++ module/protocol/sslid_replication_data_processor.h
@@ -15,8 +15,8 @@ struct sslid_replication_data_header
 {
     char virtualserver_ip[SSLID_IP_MAXLENGTH];
     unsigned short virtualserver_port;
-    unsigned int offset;
-    unsigned int size;
+    size_t offset;
+    size_t size;
 };
 
 struct sslid_replication_data
